refactor: alterados nomes de arquivos, variaves e classes genericas
