+++ .gitignore
@@ -155,3 +155,7 @@ cython_debug/
 
 # Visual Studio Code
 .vscode/
+
+# SonarQube
+.sonar/
+.scannerwork/
