+++ app/api/v1/routers/__init__.py
@@ -1,1 +1,1 @@
-SOMETHING_PREFIX = "/catalogo"
+CATALOGO_PREFIX = "/catalogo"
