--- app/api/v1/schemas/something_schema.py
@@ -1,23 +1,0 @@
-from typing import Optional
-from app.api.common.schemas import ResponseEntity, SchemaType
-
-
-class SomethingSchema(SchemaType):
-    seller_id: str
-    sku: str
-    product_name: str
-
-
-class SomethingResponse(SomethingSchema, ResponseEntity):
-    """Resposta adicionando"""
-
-
-class SomethingCreate(SomethingSchema):
-    seller_id: str
-    sku: str
-    product_name: str
-
-
-class SomethingUpdate(SchemaType):
-    """Permite apenas a atualização do nome do produto"""
-    product_name: Optional[str] = None
